move usePublicDSN to useProjectKeys hook
